backend/requirements.txt
flask
python-dotenv
openai
gunicorn
flask_cors
requests
pillow